+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - lint: Emit a warning when reporting work on an objective outside of its quarters (#306, @gpetiot)
 - Allow to report time in hours (#296, @gpetiot)
+- Improved okra-mode for generating reports (#289, @tmcgilchrist)
 
 ### Changed